requirements.txt
bcrypt==4.1.3
fastapi==0.111.0
pydantic==2.7.3
PyJWT==2.8.0
pyspark==3.5.1
pytest==8.2.2
python-decouple==3.8
sqlalchemy==2.0.30
uvicorn==0.30.1
